Roadmaps /03 Voice recognition/dataset/requirements.txt
numpy
pyserial
librosa
soundfile
tqdm
torch